Date Range Fix
There was a problem with one of the tables being joined in that was not limited to a date range so it was bringing in over 2 million records. This was outside of the scope of the problem i intended to fix but should make this query much more lightweight to run in the future. In my tests it brought the total runtime of the query down from 4+ mins to under 20 seconds